Extracted to separate module: DataSource, DataSourceGroup; added draft Stream model
